Move pipeline label to new package
