+++ common/test/python/pipeline/BUILD
@@ -1,0 +1,3 @@
+python_tests(
+    name="tests",
+)
