+++ common/test/python/users/test_authorizations.py
@@ -7,7 +7,6 @@ from pydantic import ValidationError
 from users.authorizations import (
     Activity,
     AuthMap,
-    PipelineLabel,
     StudyAuthorizations,
 )
 
@@ -207,46 +206,4 @@ class TestAuthorization:
         assert "view-form" not in authorization
 
 
-class TestPipelineLabel:
-    def test_valid(self):
-        label_object = PipelineLabel(
-            pipeline="distribution", datatype="genetic-availability", study_id="dummy"
-        )
-        label_string = label_object.model_dump()
-        assert label_string == "distribution-genetic-availability-dummy"
-        label_load = PipelineLabel.model_validate(label_string)
-        assert label_object == label_load
 
-        label_object = PipelineLabel(
-            pipeline="distribution", datatype="genetic-availability"
-        )
-        label_string = label_object.model_dump()
-        assert label_string == "distribution-genetic-availability"
-        label_load = PipelineLabel.model_validate(label_string)
-        assert label_object == label_load
-
-        label_object = PipelineLabel(
-            pipeline="distribution", datatype="form", study_id="dummy"
-        )
-        label_string = label_object.model_dump()
-        assert label_string == "distribution-form-dummy"
-        label_load = PipelineLabel.model_validate(label_string)
-        assert label_object == label_load
-
-        label_object = PipelineLabel(pipeline="distribution", datatype="form")
-        label_string = label_object.model_dump()
-        assert label_string == "distribution-form"
-        label_load = PipelineLabel.model_validate(label_string)
-        assert label_object == label_load
-
-        label_object = PipelineLabel(pipeline="accepted", study_id="dummy")
-        label_string = label_object.model_dump()
-        assert label_string == "accepted-dummy"
-        label_load = PipelineLabel.model_validate(label_string)
-        assert label_object == label_load
-
-        label_object = PipelineLabel(pipeline="accepted")
-        label_string = label_object.model_dump()
-        assert label_string == "accepted"
-        label_load = PipelineLabel.model_validate(label_string)
-        assert label_object == label_load
